requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
pydantic
crawl4ai
jinja2
